+++ src/Opaleye/Internal/Order.hs
@@ -65,20 +65,9 @@ limit' n (x, q, t) = (x, PQ.Limit (PQ.LimitOp n) q, t)
 offset' :: Int -> (a, PQ.PrimQuery, T.Tag) -> (a, PQ.PrimQuery, T.Tag)
 offset' n (x, q, t) = (x, PQ.Limit (PQ.OffsetOp n) q, t)
 
-distinctOn :: U.Unpackspec b b -> (a -> b)
-           -> (a, PQ.PrimQuery, T.Tag) -> (a, PQ.PrimQuery, T.Tag)
-distinctOn ups proj = distinctOnBy ups proj M.mempty
-
 distinctOnCorrect :: U.Unpackspec b b -> (a -> b)
                   -> (a, PQ.PrimQuery, T.Tag) -> (a, PQ.PrimQuery, T.Tag)
 distinctOnCorrect ups proj = distinctOnByCorrect ups proj M.mempty
-
-distinctOnBy :: U.Unpackspec b b -> (a -> b) -> Order a
-             -> (a, PQ.PrimQuery, T.Tag) -> (a, PQ.PrimQuery, T.Tag)
-distinctOnBy ups proj ord (cols, pq, t) = (cols, pqOut, t)
-    where pqOut = case U.collectPEs ups (proj cols) of
-            x:xs -> PQ.DistinctOnOrderBy (Just $ x NL.:| xs) (orderExprs cols ord) pq
-            []   -> pq
 
 distinctOnByCorrect :: U.Unpackspec b b -> (a -> b) -> Order a
              -> (a, PQ.PrimQuery, T.Tag) -> (a, PQ.PrimQuery, T.Tag)